Refactor: Restructure Project Directory and Modules
Details:
- Renamed module `:common` to `:core`
- Moved module `:mqtt` inside of `:common` module

Todo:
- Implement dependency injection to decouple modules. Currently, the `:logging` module depends on `:core:mqtt` as it creates the `IMessagingClient` object by directly calling `RabbitMQClient`.
