--- common/src/main/java/com/andresuryana/amlib/common/IMessagingClient.kt
@@ -1,22 +1,0 @@
-package com.andresuryana.amlib.common
-
-import com.andresuryana.amlib.common.listener.OnSubscribeTopicListener
-
-interface IMessagingClient {
-
-    @Throws(Exception::class)
-    suspend fun connect()
-
-    suspend fun publish(topic: String, payload: String): Boolean
-
-    suspend fun subscribe(topic: String, listener: OnSubscribeTopicListener)
-
-    suspend fun unsubscribe(topic: String): Boolean
-
-    @Throws(Exception::class)
-    fun disconnect()
-
-    fun isConnected(): Boolean
-
-    fun bindQueue(queue: String, routingKey: String)
-}
