--- common/src/main/java/com/andresuryana/amlib/common/data/ResourceUsage.kt
@@ -1,6 +1,0 @@
-package com.andresuryana.amlib.common.data
-
-data class ResourceUsage<T>(
-    val used: T,
-    val total: T
-)
