--- common/src/main/java/com/andresuryana/amlib/common/listener/OnSubscribeTopicListener.kt
@@ -1,6 +1,0 @@
-package com.andresuryana.amlib.common.listener
-
-interface OnSubscribeTopicListener {
-    fun onReceive(topic: String, message: String)
-    fun onCancel()
-}
