+++ core/mqtt/proguard-rules.pro
