+++ core/mqtt/src/main/AndroidManifest.xml
