+++ core/proguard-rules.pro
