+++ core/src/main/java/com/andresuryana/amlib/core/data/ResourceUsage.kt
@@ -1,0 +1,6 @@
+package com.andresuryana.amlib.core.data
+
+data class ResourceUsage<T>(
+    val used: T,
+    val total: T
+)
