+++ core/src/main/java/com/andresuryana/amlib/core/listener/OnSubscribeTopicListener.kt
@@ -1,0 +1,6 @@
+package com.andresuryana.amlib.core.listener
+
+interface OnSubscribeTopicListener {
+    fun onReceive(topic: String, message: String)
+    fun onCancel()
+}
