+++ deviceinfo/src/main/java/com/andresuryana/amlib/deviceinfo/service/DeviceInfoMonitoringService.kt
@@ -1,6 +1,6 @@
 package com.andresuryana.amlib.deviceinfo.service
 
-import com.andresuryana.amlib.common.IDeviceInfoCollector
+import com.andresuryana.amlib.core.IDeviceInfoCollector
 
 class DeviceInfoMonitoringService(private val deviceInfoCollector: IDeviceInfoCollector) {
