--- mqtt/build.gradle.kts
@@ -1,51 +1,0 @@
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.jetbrains.kotlin.android)
-}
-
-android {
-    namespace = "com.andresuryana.amlib.mqtt"
-    compileSdk = AmlibConfiguration.compileSdk
-
-    defaultConfig {
-        group = AmlibConfiguration.amlibGroup
-        version = AmlibConfiguration.amlibVersion
-
-        minSdk = AmlibConfiguration.minSdk
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-dependencies {
-
-    implementation(project(":common"))
-
-    implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.appcompat)
-    implementation(libs.material)
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.junit)
-    androidTestImplementation(libs.androidx.espresso.core)
-
-    // RabbitMQ Client
-    implementation(libs.rabbitmq.client)
-}
