+++ settings.gradle.kts
@@ -21,7 +21,7 @@ dependencyResolutionManagement {
 
 rootProject.name = "Android Monitoring Library"
 include(":app")
-include(":common")
+include(":core")
+include(":core:mqtt")
 include(":logging")
 include(":deviceinfo")
-include(":mqtt")
